added 2 containers on session details page; css applied

+++ src/views/partials/side_menu.html.tpl
@@ -48,12 +48,18 @@
                 <h1 class="side-title">Session <span class="hash">#</span>{{ session.number }}</h1>
                 <h2 class="side-subtitle"><span class="label">Run</span><span class="value">{{ run.number }}</span></h2>
             </div>
-            <div class="buttons button-actions button-actions-run">
-                <a class="button button-blue button-start {% if completed_run %}disabled{% endif%}" href="/projects/{{ session.number }}/runs/{{ run.number }}">Start</a>
-                <a class="button button-blue button-move disabled" href="/projects/{{ session.number }}/runs/{{ run.number }}">Move</a>
-                <a class="button button-blue button-continue disabled" href="/projects/{{ session.number }}/runs/{{ run.number }}">Continue</a>
-                <a class="button button-blue button-kill disabled" href="/projects/{{ session.number }}/runs/{{ run.number }}">Kill</a>
-            </div>
+            {% if not completed_run %}
+                <div class="buttons button-actions button-actions-run">
+                    <a class="button button-blue button-start {% if completed_run %}disabled{% endif%}" href="/projects/{{ session.number }}/runs/{{ run.number }}">Start</a>
+                    <a class="button button-blue button-move disabled" href="/projects/{{ session.number }}/runs/{{ run.number }}">Move</a>
+                    <a class="button button-blue button-continue disabled" href="/projects/{{ session.number }}/runs/{{ run.number }}">Continue</a>
+                    <a class="button button-blue button-kill disabled" href="/projects/{{ session.number }}/runs/{{ run.number }}">Kill</a>
+                </div>
+            {% else %}
+                <div class="buttons button-actions">
+                    <a class="button button-blue button-run button-charts" href="/projects/{{ session.number }}/runs/{{ run.number }}/stats">Stats</a>
+                </div>
+            {% endif %}
         </div>
     {% endif %}
 </div>
